refactor model families into explicit packages
