+++ gpn/msa/embedding.py
@@ -5,7 +5,7 @@ from transformers import AutoModel
 
 from gpn import register_auto_classes
 
-register_auto_classes("gpn")
+register_auto_classes("msa")
 
 
 class ModelCenterEmbedding(torch.nn.Module):
