+++ gpn/msa/vep_delta_embed.py
@@ -6,7 +6,7 @@ from transformers import AutoModel
 from gpn import register_auto_classes
 from gpn.data import Tokenizer, ReverseComplementer
 
-register_auto_classes("gpn")
+register_auto_classes("msa")
 
 
 class VEPDeltaEmbed(torch.nn.Module):
