+++ gpn/msa/vep_influence.py
@@ -7,7 +7,7 @@ from transformers import AutoModelForMaskedLM
 from gpn import register_auto_classes
 from gpn.data import Tokenizer, ReverseComplementer
 
-register_auto_classes("gpn")
+register_auto_classes("msa")
 
 
 class VEPInfluence(torch.nn.Module):
