+++ gpn/ss/get_logits.py
@@ -13,7 +13,7 @@ from transformers import AutoTokenizer, AutoModelForMaskedLM, Trainer, TrainingA
 from gpn import register_auto_classes
 from gpn.data import Genome, load_dataset_from_file_or_dir, token_input_id
 
-register_auto_classes("gpn")
+register_auto_classes("ss")
 
 
 class MLMforLogitsModel(torch.nn.Module):
