+++ gpn/ss/run_vep_influence.py
@@ -14,7 +14,7 @@ from transformers import AutoTokenizer, AutoModelForMaskedLM, Trainer, TrainingA
 from gpn import register_auto_classes
 from gpn.data import Genome, load_dataset_from_file_or_dir, token_input_id
 
-register_auto_classes("gpn")
+register_auto_classes("ss")
 
 
 class ModelforVEPModel(torch.nn.Module):
