+++ gpn/ss/run_vep_mean_euclidean_dist.py
@@ -14,7 +14,7 @@ from transformers import AutoTokenizer, AutoModel, Trainer, TrainingArguments
 from gpn import register_auto_classes
 from gpn.data import Genome, load_dataset_from_file_or_dir, token_input_id
 
-register_auto_classes("gpn")
+register_auto_classes("ss")
 
 
 class ModelforVEPModel(torch.nn.Module):
