+++ tests/test_registration.py
@@ -39,8 +39,8 @@ assert "transformers" not in sys.modules
     assert result.returncode == 0, result.stderr
 
 
-def test_gpn_registration_is_explicit_and_idempotent():
-    from gpn.model import (
+def test_ss_registration_is_explicit_and_idempotent():
+    from gpn.ss.model import (
         GPNConfig,
         GPNForMaskedLM,
         GPNForSequenceClassification,
@@ -48,8 +48,8 @@ def test_gpn_registration_is_explicit_and_idempotent():
         GPNModel,
     )
 
-    register_auto_classes("gpn")
-    register_auto_classes("gpn")
+    register_auto_classes("ss")
+    register_auto_classes("ss")
 
     assert type(AutoConfig.for_model("GPN")) is GPNConfig
     assert AutoModel._model_mapping[GPNConfig] is GPNModel
@@ -75,11 +75,11 @@ def test_star_registration_is_explicit_and_idempotent():
     assert AutoModelForMaskedLM._model_mapping[GPNStarConfig] is GPNStarForMaskedLM
 
 
-def test_phylogpn_registration_is_explicit_and_idempotent(tmp_path: Path):
-    from gpn.phylogpn import PhyloGPNConfig, PhyloGPNModel, PhyloGPNTokenizer
+def test_phylo_registration_is_explicit_and_idempotent(tmp_path: Path):
+    from gpn.phylo.model import PhyloGPNConfig, PhyloGPNModel, PhyloGPNTokenizer
 
-    register_auto_classes("phylogpn")
-    register_auto_classes("phylogpn")
+    register_auto_classes("phylo")
+    register_auto_classes("phylo")
 
     config = AutoConfig.for_model(
         "phylogpn",
@@ -114,7 +114,7 @@ def test_phylogpn_registration_is_explicit_and_idempotent(tmp_path: Path):
 
 
 def test_tiny_gpn_model_round_trip(tmp_path: Path):
-    register_auto_classes("gpn")
+    register_auto_classes("ss")
 
     config = AutoConfig.for_model(
         "GPN",
@@ -141,8 +141,9 @@ def test_tiny_gpn_model_round_trip(tmp_path: Path):
     torch.testing.assert_close(actual, expected)
 
 
-def test_published_gpn_architectures_construct_and_run_with_auto_classes():
-    register_auto_classes("gpn")
+def test_published_ss_and_msa_architectures_construct_with_auto_classes():
+    register_auto_classes("ss")
+    register_auto_classes("msa")
 
     convnet_config = AutoConfig.for_model(
         "ConvNet",
@@ -170,6 +171,33 @@ def test_published_gpn_architectures_construct_and_run_with_auto_classes():
     with torch.no_grad():
         roformer_logits = roformer(input_ids=input_ids[:, :6]).logits
     assert roformer_logits.shape == (1, 6, 6)
+
+
+def test_registering_msa_does_not_register_ss():
+    """Family selectors should not import or register unrelated implementations."""
+
+    code = """
+import sys
+
+from transformers import AutoConfig
+from gpn import register_auto_classes
+
+register_auto_classes("msa")
+assert type(AutoConfig.for_model("GPNRoFormer")).__name__ == "GPNMSAConfig"
+assert "gpn.ss.model" not in sys.modules
+"""
+    result = subprocess.run(
+        [sys.executable, "-c", code],
+        check=False,
+        capture_output=True,
+        env={
+            **os.environ,
+            "HF_HUB_OFFLINE": "1",
+            "TRANSFORMERS_OFFLINE": "1",
+        },
+        text=True,
+    )
+    assert result.returncode == 0, result.stderr
 
 
 def test_unknown_registration_family_is_rejected():
